# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servidor API para el Proyecto Sakoo (Arquitectura Limpia).",
        "title": "Sakoo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte Sakoo",
            "url": "http://www.sakoo.com",
            "email": "soporte@sakoo.com"
        },
        "license": {
            "name": "Propietaria",
            "url": "http://www.sakoo.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "Autentica a un usuario mediante su correo electrónico y contraseña (cifrada con RSA en tránsito) y devuelve un token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "parameters": [
                    {
                        "description": "Credenciales de acceso",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesión iniciada correctamente",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-github_com_aaron_sakoo-backend_internal_domain_AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Elimina de forma lógica la cuenta del usuario autenticado validando su correo electrónico y contraseña (cifrada en tránsito).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Eliminar cuenta (Legacy)",
                "parameters": [
                    {
                        "description": "Credenciales de confirmación de eliminación",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteLegacyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada lógicamente de manera exitosa",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/auth/register": {
            "post": {
                "description": "Registra un nuevo usuario en la plataforma validando el código OTP y descifrando la contraseña.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_domain.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al registrar usuario o datos duplicados",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/account": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Elimina de forma lógica la cuenta del usuario autenticado validando un código OTP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Eliminar cuenta (v1)",
                "parameters": [
                    {
                        "description": "Confirmación de eliminación con código OTP",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada lógicamente de manera exitosa",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "400": {
                        "description": "Código OTP inválido o error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/otp/request": {
            "post": {
                "description": "Genera y envía un código OTP por correo electrónico al usuario para registrarse, recuperar contraseña o eliminar cuenta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Solicitar OTP",
                "parameters": [
                    {
                        "description": "Cuerpo de la solicitud con email y acción",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.RequestOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Código OTP generado y enviado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password/reset": {
            "post": {
                "description": "Restablece la contraseña de un usuario validando el código OTP enviado a su correo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Restablecer contraseña",
                "parameters": [
                    {
                        "description": "Datos para restablecer la contraseña",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contraseña restablecida correctamente",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "400": {
                        "description": "Código OTP inválido o credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Registra un nuevo usuario en la plataforma validando el código OTP y descifrando la contraseña.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_domain.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al registrar usuario o datos duplicados",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna los datos de perfil del usuario actualmente autenticado (mediante JWT).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener perfil del usuario",
                "responses": {
                    "200": {
                        "description": "Perfil del usuario recuperado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-internal_api_ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "track_code": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aaron_sakoo-backend_internal_api_response.APIResponse-github_com_aaron_sakoo-backend_internal_domain_AuthResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_domain.AuthResponse"
                },
                "message": {
                    "type": "string"
                },
                "track_code": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aaron_sakoo-backend_internal_domain.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aaron_sakoo-backend_internal_domain.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aaron_sakoo-backend_internal_domain.RegisterRequest": {
            "type": "object",
            "properties": {
                "document_number": {
                    "type": "string"
                },
                "document_type_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "otp_code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "second_last_name": {
                    "type": "string"
                },
                "user_type_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.DeleteAccountRequest": {
            "type": "object",
            "properties": {
                "otp_code": {
                    "type": "string"
                }
            }
        },
        "internal_api.DeleteLegacyAccountRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Coloque el token JWT de la siguiente manera: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}